Add Spring Boot Web dependency to pom.xml
The Spring Boot Starter Web dependency was added to the pom.xml file. This was done to support the implementation of a new REST API endpoint and improve code organization with the creation of new TableConfiguration and TableDrivenAgent classes. The Lombok annotation processor is also included for reducing boilerplate code.